Starting Layout for combining the tools
